[PRDP-17] Changed packages' name and renamed Client classes
