--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/DebtorPosition.java
@@ -1,14 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class DebtorPosition {
-	private String modelType;
-	private String noticeNumber;
-	private String iuv;
-}
