--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/MapEntry.java
@@ -1,15 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class MapEntry {
-	private String key;    
-	private String value;
-}
