--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/Transaction.java
@@ -1,23 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Transaction {
-	private long idTransaction;
-	private long grandTotal;
-	private long amount;
-	private long fee;
-	private String transactionStatus;
-	private String accountingStatus;
-	private String rrn;
-	private String authorizationCode;
-	private String creationDate;
-	private String numAut;
-	private String accountCode;
-	private TransactionPsp psp;
-}
