--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/WalletItem.java
@@ -1,23 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event;
-
-import it.gov.pagopa.receipt.pdf.datastore.entities.event.enumeration.WalletType;
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class WalletItem {
-	private String idWallet;
-	private WalletType walletType;
-	private List<String> enableableFunctions;
-	private boolean pagoPa;
-	private String onboardingChannel;
-	private boolean favourite;
-	private String createDate;
-	private Info info;
-	private AuthRequest authRequest;
-}
