--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/enumeration/BizEventStatusType.java
@@ -1,5 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event.enumeration;
-
-public enum BizEventStatusType {
-    NA, RETRY, FAILED, DONE
-}
