--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/event/enumeration/WalletType.java
@@ -1,5 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.event.enumeration;
-
-public enum WalletType {
-	CARD, PAYPAL, BANCOMATPAY
-}
