--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/receipt/IOMessageData.java
@@ -1,13 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.receipt;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class IOMessageData {
-    private String idMessageDebtor;
-    private String idMessagePayer;
-}
