--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/receipt/ReceiptMetadata.java
@@ -1,14 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.receipt;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class ReceiptMetadata {
-
-    private String name;
-    private String url;
-}
