--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/receipt/enumeration/ReasonErrorCode.java
@@ -1,19 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.receipt.enumeration;
-
-public enum ReasonErrorCode {
-    ERROR_QUEUE(902), ERROR_BLOB_STORAGE(901), ERROR_PDF_ENGINE(0);
-
-    private int code;
-
-    ReasonErrorCode(int code){
-        this.code = code;
-    }
-
-    public int getCode(){
-        return this.code;
-    }
-
-    public int getCustomCode(int customCode){
-        return customCode;
-    }
-}
