--- src/main/java/it/gov/pagopa/receipt/pdf/datastore/entities/receipt/enumeration/ReceiptStatusType.java
@@ -1,5 +1,0 @@
-package it.gov.pagopa.receipt.pdf.datastore.entities.receipt.enumeration;
-
-public enum ReceiptStatusType {
-    NOT_QUEUE_SENT, INSERTED, RETRY, GENERATED, SIGNED, FAILED, IO_NOTIFIED
-}
