+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/event/TransactionDetails.java
@@ -1,0 +1,15 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.event;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TransactionDetails {
+	private String origin;
+	private User user;
+	private Transaction transaction;
+	private WalletItem wallet;
+}
