+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/event/TransactionPsp.java
@@ -1,0 +1,14 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.event;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TransactionPsp {
+	private String idChannel;
+	private String businessName;
+	private String serviceName;
+}
