+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/event/enumeration/BizEventStatusType.java
@@ -1,0 +1,5 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.event.enumeration;
+
+public enum BizEventStatusType {
+    NA, RETRY, FAILED, DONE
+}
