+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/receipt/IOMessageData.java
@@ -1,0 +1,13 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.receipt;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class IOMessageData {
+    private String idMessageDebtor;
+    private String idMessagePayer;
+}
