+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/receipt/ReasonError.java
@@ -1,0 +1,16 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.receipt;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReasonError {
+    private int code;
+    private String message;
+
+}
