+++ src/main/java/it/gov/pagopa/receipt/pdf/datastore/entity/receipt/enumeration/ReceiptStatusType.java
@@ -1,0 +1,5 @@
+package it.gov.pagopa.receipt.pdf.datastore.entity.receipt.enumeration;
+
+public enum ReceiptStatusType {
+    NOT_QUEUE_SENT, INSERTED, RETRY, GENERATED, SIGNED, FAILED, IO_NOTIFIED
+}
